Rename admin apis
