+++ backend/src/main/kotlin/fr/gouv/cnsp/monitorfish/infrastructure/api/bff/FleetSegmentController.kt
@@ -1,13 +1,12 @@
 package fr.gouv.cnsp.monitorfish.infrastructure.api.bff
 
-import fr.gouv.cnsp.monitorfish.domain.use_cases.fleet_segment.*
-import fr.gouv.cnsp.monitorfish.infrastructure.api.input.CreateOrUpdateFleetSegmentDataInput
+import fr.gouv.cnsp.monitorfish.domain.use_cases.fleet_segment.ComputeFleetSegments
+import fr.gouv.cnsp.monitorfish.domain.use_cases.fleet_segment.GetAllFleetSegmentsByYear
 import fr.gouv.cnsp.monitorfish.infrastructure.api.outputs.FleetSegmentDataOutput
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
 import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.websocket.server.PathParam
-import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.*
 
 @RestController
@@ -15,11 +14,6 @@ import org.springframework.web.bind.annotation.*
 @Tag(name = "APIs for Fleet segments")
 class FleetSegmentController(
     private val getAllFleetSegmentsByYearByYear: GetAllFleetSegmentsByYear,
-    private val updateFleetSegment: UpdateFleetSegment,
-    private val deleteFleetSegment: DeleteFleetSegment,
-    private val createFleetSegment: CreateFleetSegment,
-    private val getFleetSegmentYearEntries: GetFleetSegmentYearEntries,
-    private val addFleetSegmentYear: AddFleetSegmentYear,
     private val computeFleetSegments: ComputeFleetSegments,
 ) {
 
@@ -33,71 +27,6 @@ class FleetSegmentController(
         return getAllFleetSegmentsByYearByYear.execute(year).map { fleetSegment ->
             FleetSegmentDataOutput.fromFleetSegment(fleetSegment)
         }
-    }
-
-    @PutMapping(value = ["/backoffice"], consumes = ["application/json"])
-    @Operation(summary = "Update a fleet segment")
-    fun updateFleetSegment(
-        @Parameter(description = "Year")
-        @RequestParam(name = "year")
-        year: Int,
-        @Parameter(description = "Segment")
-        @RequestParam(name = "segment")
-        segment: String,
-        @RequestBody
-        createOrUpdateFleetSegmentData: CreateOrUpdateFleetSegmentDataInput,
-    ): FleetSegmentDataOutput {
-        val updatedFleetSegment = updateFleetSegment.execute(
-            segment = segment,
-            fields = createOrUpdateFleetSegmentData.toCreateOrUpdateFleetSegmentFields(),
-            year = year,
-        )
-
-        return FleetSegmentDataOutput.fromFleetSegment(updatedFleetSegment)
-    }
-
-    @DeleteMapping(value = ["/backoffice"])
-    @Operation(summary = "Delete a fleet segment")
-    fun deleteFleetSegment(
-        @Parameter(description = "Year")
-        @RequestParam(name = "year")
-        year: Int,
-        @Parameter(description = "Segment")
-        @RequestParam(name = "segment")
-        segment: String,
-    ): List<FleetSegmentDataOutput> {
-        return deleteFleetSegment.execute(segment, year).map {
-            FleetSegmentDataOutput.fromFleetSegment(it)
-        }
-    }
-
-    @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping(value = ["/backoffice"])
-    @Operation(summary = "Create a fleet segment")
-    fun createFleetSegment(
-        @RequestBody
-        newFleetSegmentData: CreateOrUpdateFleetSegmentDataInput,
-    ): FleetSegmentDataOutput {
-        val createdFleetSegment = createFleetSegment.execute(newFleetSegmentData.toCreateOrUpdateFleetSegmentFields())
-
-        return FleetSegmentDataOutput.fromFleetSegment(createdFleetSegment)
-    }
-
-    @GetMapping("/backoffice/years")
-    @Operation(summary = "Get fleet segment year entries")
-    fun getFleetSegmentYearEntries(): List<Int> {
-        return getFleetSegmentYearEntries.execute()
-    }
-
-    @ResponseStatus(HttpStatus.CREATED)
-    @PostMapping("/backoffice/{year}")
-    @Operation(summary = "Add a fleet segment year")
-    fun addFleetSegmentYear(
-        @PathParam("Year")
-        @PathVariable(name = "year")
-        year: Int,
-    ) {
-        return addFleetSegmentYear.execute(year)
     }
 
     @GetMapping("/compute")
