+++ frontend/cypress/e2e/backoffice/control_objectives.spec.ts
@@ -4,7 +4,7 @@ context('Control objectives', () => {
   beforeEach(() => {
     const currentYear = new Date().getFullYear()
     cy.intercept('GET', `/bff/v1/fleet_segments/${currentYear}`).as('fleetSegments')
-    cy.intercept('GET', `/bff/v1/control_objectives/${currentYear}`).as('controlObjectives')
+    cy.intercept('GET', `/bff/v1/admin/control_objectives/${currentYear}`).as('controlObjectives')
     cy.visit('/backoffice/control_objectives')
     cy.wait('@fleetSegments')
     cy.wait('@controlObjectives')
@@ -45,7 +45,7 @@ context('Control objectives', () => {
 
   it('Should update the targetNumberOfControlsAtPort field on an objective', () => {
     // When
-    cy.intercept('PUT', '/bff/v1/control_objectives/78').as('updateObjective')
+    cy.intercept('PUT', '/bff/v1/admin/control_objectives/78').as('updateObjective')
     cy.wait(50)
     cy.get('*[data-cy="row-78-targetNumberOfControlsAtPort"]')
       .type('{backspace}{backspace}{backspace}{backspace}{backspace}')
@@ -58,7 +58,7 @@ context('Control objectives', () => {
     cy.get('*[data-cy="row-78-targetNumberOfControlsAtPort"]').should('have.value', '23')
 
     // The value is saved in database when I refresh the page
-    cy.intercept('GET', '/bff/v1/control_objectives').as('controlObjectives')
+    cy.intercept('GET', '/bff/v1/admin/control_objectives').as('controlObjectives')
     cy.visit('/backoffice/control_objectives')
     cy.wait('@controlObjectives')
     cy.wait(50)
@@ -67,7 +67,7 @@ context('Control objectives', () => {
 
   it('Should update the targetNumberOfControlsAtSea field on an objective', () => {
     // When
-    cy.intercept('PUT', '/bff/v1/control_objectives/78').as('updateObjective')
+    cy.intercept('PUT', '/bff/v1/admin/control_objectives/78').as('updateObjective')
     cy.wait(50)
     cy.get('*[data-cy="row-78-targetNumberOfControlsAtSea"]')
       .type('{backspace}{backspace}{backspace}{backspace}{backspace}')
@@ -80,7 +80,7 @@ context('Control objectives', () => {
     cy.get('*[data-cy="row-78-targetNumberOfControlsAtSea"]').should('have.value', '23')
 
     // The value is saved in database when I refresh the page
-    cy.intercept('GET', '/bff/v1/control_objectives').as('controlObjectives')
+    cy.intercept('GET', '/bff/v1/admin/control_objectives').as('controlObjectives')
     cy.visit('/backoffice/control_objectives')
     cy.wait('@controlObjectives')
     cy.wait(50)
@@ -89,7 +89,7 @@ context('Control objectives', () => {
 
   it('Should update the controlPriorityLevel field on an objective', () => {
     // When
-    cy.intercept('PUT', '/bff/v1/control_objectives/78').as('updateObjective')
+    cy.intercept('PUT', '/bff/v1/admin/control_objectives/78').as('updateObjective')
     cy.wait(50)
     cy.get('[data-cy="row-78-controlPriorityLevel"]').parent().click()
     cy.get('.rs-picker-select-menu-item').eq(2).click()
@@ -100,7 +100,7 @@ context('Control objectives', () => {
     cy.get('[data-cy="row-78-controlPriorityLevel"]').should('exist')
 
     // The value is saved in database when I refresh the page
-    cy.intercept('GET', '/bff/v1/control_objectives').as('controlObjectives')
+    cy.intercept('GET', '/bff/v1/admin/control_objectives').as('controlObjectives')
     cy.visit('/backoffice/control_objectives')
     cy.wait('@controlObjectives')
     cy.wait(50)
@@ -110,7 +110,7 @@ context('Control objectives', () => {
   it('Should delete an objective', () => {
     // Given
     cy.get('.rs-table-row').should('have.length', 67)
-    cy.intercept('DELETE', '/bff/v1/control_objectives/78').as('deleteObjective')
+    cy.intercept('DELETE', '/bff/v1/admin/control_objectives/78').as('deleteObjective')
 
     // When
     cy.get('*[data-cy="delete-row-78"]').click()
@@ -121,7 +121,7 @@ context('Control objectives', () => {
     cy.get('.rs-table-row').should('have.length', 66)
 
     // The value is saved in database when I refresh the page
-    cy.intercept('GET', '/bff/v1/control_objectives').as('controlObjectives')
+    cy.intercept('GET', '/bff/v1/admin/control_objectives').as('controlObjectives')
     cy.visit('/backoffice/control_objectives')
     cy.wait('@controlObjectives')
     cy.wait(50)
@@ -132,7 +132,7 @@ context('Control objectives', () => {
   it('Should add an objective', () => {
     // Given
     cy.get('.rs-table-row').should('have.length', 66)
-    cy.intercept('POST', '/bff/v1/control_objectives').as('addObjective')
+    cy.intercept('POST', '/bff/v1/admin/control_objectives').as('addObjective')
     cy.wait(200)
 
     // When
@@ -145,7 +145,7 @@ context('Control objectives', () => {
     cy.get('.rs-table-row').should('have.length', 67)
 
     // Update the row when the value is updated in local memory
-    cy.intercept('PUT', '/bff/v1/control_objectives/125').as('updateObjective')
+    cy.intercept('PUT', '/bff/v1/admin/control_objectives/125').as('updateObjective')
     cy.get('[data-cy="row-125-targetNumberOfControlsAtPort"]').type(
       '{backspace}{backspace}{backspace}{backspace}{backspace}'
     )
@@ -155,7 +155,7 @@ context('Control objectives', () => {
     cy.get('[data-cy="row-125-targetNumberOfControlsAtPort"]').should('exist')
 
     // The value is saved in database when I refresh the page
-    cy.intercept('GET', '/bff/v1/control_objectives').as('controlObjectives')
+    cy.intercept('GET', '/bff/v1/admin/control_objectives').as('controlObjectives')
     cy.visit('/backoffice/control_objectives')
     cy.wait('@controlObjectives')
     cy.wait(50)
@@ -183,7 +183,7 @@ context('Control objectives', () => {
     cy.get('.rs-table-row').should('have.length', 67)
     cy.get('*[data-cy^="control-objectives-year"]').click()
     cy.get('.rs-picker-select-menu-item').should('have.length', 2)
-    cy.intercept('POST', '/bff/v1/control_objectives/years').as('addObjectiveYear')
+    cy.intercept('POST', '/bff/v1/admin/control_objectives/years').as('addObjectiveYear')
 
     // When
     cy.get('*[data-cy="control-objectives-add-year"]').click()
