+++ frontend/cypress/e2e/backoffice/fleet_segments.spec.ts
@@ -35,7 +35,7 @@ context('Fleet segments', () => {
     cy.log('Should update the segment')
 
     // When
-    cy.intercept('PUT', `/bff/v1/fleet_segments/backoffice?year=${currentYear}&segment=ATL01`).as('updateFleetSegment')
+    cy.intercept('PUT', `/bff/v1/admin/fleet_segments/?year=${currentYear}&segment=ATL01`).as('updateFleetSegment')
 
     cy.get('[aria-rowindex="2"]').find('[title="Editer la ligne"]').click()
     cy.fill('Nom', 'ATL036')
@@ -92,7 +92,7 @@ context('Fleet segments', () => {
 
   it('Should create a new fleet segment and delete it', () => {
     // Given
-    cy.intercept('POST', '/bff/v1/fleet_segments/backoffice').as('createFleetSegment')
+    cy.intercept('POST', '/bff/v1/admin/fleet_segments').as('createFleetSegment')
 
     // When
     cy.clickButton('Ajouter un segment')
@@ -137,9 +137,7 @@ context('Fleet segments', () => {
     // -------------------------------------------------------------------------
     cy.log('Should delete a fleet segment')
 
-    cy.intercept('DELETE', `/bff/v1/fleet_segments/backoffice?year=${currentYear}&segment=ABC123`).as(
-      'deleteFleetSegment'
-    )
+    cy.intercept('DELETE', `/bff/v1/admin/fleet_segments/?year=${currentYear}&segment=ABC123`).as('deleteFleetSegment')
     cy.get('[aria-rowindex="2"]').find('[title="Supprimer la ligne"]').click()
     cy.wait('@deleteFleetSegment')
